feat: api 구현 마무리

+++ controller/CartController.js
@@ -1,9 +1,7 @@
 const jwt = require("jsonwebtoken");
 const conn = require("../mariadb");
 const { StatusCodes } = require("http-status-codes");
-const dotenv = require("dotenv");
-
-dotenv.config();
+const ensureAuthorization = require("../auth");
 
 // 장바구니 담기
 const addToCart = (req, res) => {
@@ -51,9 +49,15 @@ const getCartItems = (req, res) => {
     let sql = `SELECT cartItems.id, book_id, title, summary, quantity, price 
                 FROM cartItems LEFT JOIN books 
                 ON cartItems.book_id = books.id
-                WHERE user_id=? AND cartItems.id IN (?)`;
+                WHERE user_id=?`;
 
-    let values = [authorization.id, selected];
+    let values = [authorization.id];
+
+    if (selected) {
+      // 주문서 작성 시 선택한 장바구니 목록 조회
+      sql += ` AND cartItems.id IN (?)`;
+      values.push(selected);
+    }
 
     conn.query(sql, values, (err, results) => {
       if (err) {
@@ -67,31 +71,30 @@ const getCartItems = (req, res) => {
 };
 
 const removeCartItem = (req, res) => {
-  const cartItemId = req.params.id;
+  let authorization = ensureAuthorization(req);
 
-  let sql = "DELETE FROM cartItems WHERE id=?;";
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요.",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "잘못된 토큰입니다.",
+    });
+  } else {
+    const cartItemId = req.params.id;
 
-  conn.query(sql, cartItemId, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
+    let sql = "DELETE FROM cartItems WHERE id=?;";
 
-    return res.status(StatusCodes.OK).json(results);
-  });
-};
+    conn.query(sql, cartItemId, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
 
-function ensureAuthorization(req) {
-  try {
-    let receivedJwt = req.headers["authorization"];
-    let decodedJwt = jwt.verify(receivedJwt, process.env.SECRET_KEY);
-    return decodedJwt;
-  } catch (err) {
-    console.log(err.name);
-    console.log(err.message);
-
-    return err;
+      return res.status(StatusCodes.OK).json(results);
+    });
   }
-}
+};
 
 module.exports = { addToCart, getCartItems, removeCartItem };
