+++ controller/OrderController.js
@@ -1,6 +1,8 @@
 // const conn = require("../mariadb");
 const mariadb = require("mysql2/promise");
 const { StatusCodes } = require("http-status-codes");
+const jwt = require("jsonwebtoken");
+const ensureAuthorization = require("../auth");
 
 const order = async (req, res) => {
   const conn = await mariadb.createConnection({
@@ -11,42 +13,60 @@ const order = async (req, res) => {
     dataStrings: true,
   });
 
-  const { items, delivery, totalQuantity, totalPrice, userId, firstBookTitle } =
-    req.body;
+  let authorization = ensureAuthorization(req);
 
-  // delivery
-  let sql =
-    "INSERT INTO delivery (address, receiver, contact) VALUES (?, ?, ?);";
-  let values = [delivery.address, delivery.receiver, delivery.contact];
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요.",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "잘못된 토큰입니다.",
+    });
+  } else {
+    const { items, delivery, totalQuantity, totalPrice, firstBookTitle } =
+      req.body;
 
-  let [results] = await conn.execute(sql, values);
-  let delivery_id = results.insertId;
+    // delivery
+    let sql =
+      "INSERT INTO delivery (address, receiver, contact) VALUES (?, ?, ?);";
+    let values = [delivery.address, delivery.receiver, delivery.contact];
 
-  // orders
-  sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id)
+    let [results] = await conn.execute(sql, values);
+    let delivery_id = results.insertId;
+
+    // orders
+    sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id)
             VALUES (?, ?, ?, ?, ?);`;
-  values = [firstBookTitle, totalQuantity, totalPrice, userId, delivery_id];
+    values = [
+      firstBookTitle,
+      totalQuantity,
+      totalPrice,
+      authorization.id,
+      delivery_id,
+    ];
 
-  [results] = await conn.execute(sql, values);
-  let order_id = results.insertId;
+    [results] = await conn.execute(sql, values);
+    let order_id = results.insertId;
 
-  // items를 가지고 장바구니에서 book_id, quantity를 조회
-  sql = `SELECT book_id, quantity FROM cartItems WHERE id IN (?);`;
-  let [orderItems, fields] = await conn.query(sql, [items]);
+    // items를 가지고 장바구니에서 book_id, quantity를 조회
+    sql = `SELECT book_id, quantity FROM cartItems WHERE id IN (?);`;
+    let [orderItems, fields] = await conn.query(sql, [items]);
 
-  // orderedBook
-  sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?;`;
-  values = [];
+    // orderedBook
+    sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?;`;
+    values = [];
 
-  // items가 배열로 넘어오고 이 안에서 요소들을 하나씩 꺼내서 forEach문을 돌림
-  orderItems.forEach((item) => {
-    values.push([order_id, item.book_id, item.quantity]);
-  });
-  results = await conn.query(sql, [values]);
+    // items가 배열로 넘어오고 이 안에서 요소들을 하나씩 꺼내서 forEach문을 돌림
+    orderItems.forEach((item) => {
+      values.push([order_id, item.book_id, item.quantity]);
+    });
+    results = await conn.query(sql, [values]);
 
-  let result = await deleteCartItems(conn, items);
+    let result = await deleteCartItems(conn, items);
 
-  return res.status(StatusCodes.OK).json(result);
+    return res.status(StatusCodes.OK).json(result);
+  }
 };
 
 const deleteCartItems = async (conn, items) => {
@@ -57,41 +77,65 @@ const deleteCartItems = async (conn, items) => {
 };
 
 const getOrders = async (req, res) => {
-  const conn = await mariadb.createConnection({
-    host: "127.0.0.1",
-    user: "root",
-    password: "root",
-    database: "Bookshop",
-    dataStrings: true,
-  });
+  let authorization = ensureAuthorization(req);
 
-  let sql = `SELECT orders.id, created_at, address, receiver, contact, 
-            book_title, total_quantity, total_price   
-            FROM orders LEFT JOIN delivery 
-            ON orders.delivery_id = delivery.id`;
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요.",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "잘못된 토큰입니다.",
+    });
+  } else {
+    const conn = await mariadb.createConnection({
+      host: "127.0.0.1",
+      user: "root",
+      password: "root",
+      database: "Bookshop",
+      dataStrings: true,
+    });
 
-  let [rows, fields] = await conn.query(sql);
-  return res.status(StatusCodes.OK).json(rows);
+    let sql = `SELECT orders.id, created_at, address, receiver, contact, 
+              book_title, total_quantity, total_price   
+              FROM orders LEFT JOIN delivery 
+              ON orders.delivery_id = delivery.id`;
+
+    let [rows, fields] = await conn.query(sql);
+    return res.status(StatusCodes.OK).json(rows);
+  }
 };
 
 const getOrderDetail = async (req, res) => {
-  const conn = await mariadb.createConnection({
-    host: "127.0.0.1",
-    user: "root",
-    password: "root",
-    database: "Bookshop",
-    dataStrings: true,
-  });
+  let authorization = ensureAuthorization(req);
 
-  const { id } = req.params;
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요.",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "잘못된 토큰입니다.",
+    });
+  } else {
+    const conn = await mariadb.createConnection({
+      host: "127.0.0.1",
+      user: "root",
+      password: "root",
+      database: "Bookshop",
+      dataStrings: true,
+    });
 
-  let sql = `SELECT book_id, title, author, price, quantity 
-            FROM orderedBook LEFT JOIN books 
-            ON orderedBook.book_id = books.id
-            WHERE order_id = ?`;
+    const orderId = req.params.id;
 
-  let [rows, fields] = await conn.query(sql, [id]);
-  return res.status(StatusCodes.OK).json(rows);
+    let sql = `SELECT book_id, title, author, price, quantity 
+              FROM orderedBook LEFT JOIN books 
+              ON orderedBook.book_id = books.id
+              WHERE order_id = ?`;
+
+    let [rows, fields] = await conn.query(sql, [orderId]);
+    return res.status(StatusCodes.OK).json(rows);
+  }
 };
 
 module.exports = {
